4.17.  функции сортировки
