+++ src/presenter/event-list-presenter.js
@@ -6,7 +6,8 @@ import SortView from '../view/sort-view.js';
 import ListEmptyView from '../view/list-empty-view.js';
 import { FILTER_TYPES } from '../constants.js';
 import FilterView from '../view/filter-view.js';
-import { filterEvents, generateFilters } from '../utils/utils.js';
+// import { filterEvents, generateFilters } from '../utils/utils.js';
+import { filterEvents, generateFilters } from '../utils/filtr-event.js';
 export default class EventListPresenter {
   #eventListContainer = null;
   #tripFiltersContainer = null;
