+++ src/utils/utils.js
@@ -2,8 +2,6 @@ import dayjs from 'dayjs';
 import { DateFormat } from '../constants';
 import duration from 'dayjs/plugin/duration';
 dayjs.extend(duration);
-import {FilterType, SortType, DATE_NOW } from '../constants';
-import EventsModel from '../model/events-model';
 
 
 const getRandomArrayElement = (items) =>items[Math.floor(Math.random() * items.length)];
@@ -45,35 +43,26 @@ const getDuration = (dateFrom, dateTo) => {
 };
 
 
-// const eventsDate = new EventsModel();
+// import EventsModel from '../model/events-model';
+// import { SortType } from '../constants';
 
-// const filterEventEverything = (events) => events;
-// const filterEventFuture = (events) => events.filter((event) => event.dateFrom > DATE_NOW);
-// const filterEventPresent = (events) => events.filter((event) => event.dateFrom <= DATE_NOW && event.dateTo >= DATE_NOW);
-// const filterEventPast = (events) => events.filter((event) => event.dateTo < DATE_NOW);
-
-const filterEvents = {
-  [FilterType.EVERYTHING]: (events) => events,
-  [FilterType.FUTURE]: (events) => events.filter((item) => Date.parse(item.dateFrom) > Date.parse(DATE_NOW)),
-  [FilterType.PRESENT]: (events) => events.filter((item) => Date.parse(item.dateFrom) <= Date.parse(DATE_NOW) && Date.parse(item.dateTo) >= Date.parse(DATE_NOW)),
-  [FilterType.PAST]: (events) => events.filter((item) => Date.parse(item.dateTo) < Date.parse(DATE_NOW)),
-};
-function generateFilters(events) {
-  return Object.entries(filterEvents).map(([filterType, filterTask]) => ({
-    type: filterType,
-    count: filterTask(events).length,
-  }));
-}
-// console.log(generateFilters(eventsDate.events));
+// const event = new EventsModel;
 // const sortEvents = {
-//   [SortType.Day]: (events) => events,
-//   [SortType.TIME]: (events) => events.sort((a, b) => return new Date.parse(item.dateFrom) > Date.parse(DATE_NOW)),
-//   [SortType.PRICE]: (events) => events.sort((item) => Date.parse(item.price) <= Date.parse(DATE_NOW)),
+//   [SortType.DAY]: (events) => events.sort((a, b) => new Date(b.dateFrom) - new Date(a.dateFrom)),
+//   [SortType.EVENT]: (events) => events.sort((a, b) => a.type.localeCompare(b.type)),
+//   [SortType.PRICE]: (events) => events.sort((a, b) => b.basePrice - a.basePrice),
+//   [SortType.TIME]: (events) => events.sort((a, b) => new Date(b.dateTo - b.dateFrom) - new Date(a.dateTo - a.dateFrom)),
+//   [SortType.TYPE]: (events) => events.sort((a, b) => b.offers.length - a.offers.length),
+// };
+// function generateSort(events) {
+//   return Object.entries(sortEvents).map(([sortType, sortTask]) => ({
+//     type: sortType,
+//     count: sortTask(events).length,
+//   }));
 // }
-// console.log(DATE_NOW);
-// console.log(eventsDate);
-// console.log(filterEvents['past'](eventsDate.events));
 
 
-export {getRandomArrayElement, getRandomInteger, getRandomDate, formatDate, formatTime, formatDateForm, getDuration, filterEvents, generateFilters};
+// console.table(generateSort(event.events));
+
+export {getRandomArrayElement, getRandomInteger, getRandomDate, formatDate, formatTime, formatDateForm, getDuration};
